Added preliminary support for colour animations. This was done by modifying the fundamental data type of a TCCube object from a boolean to a byte. Colours are handled by the TCAnim object, where each colour is represented internally by a different cube state. This allows you to use the conventional methods in the cube states to modify each colour value individually, or use the TCAnim functions to modify the colours of voxels all at once.
Currently, Lua greyscale and colour animations are NOT working. The new
animations object type works fine, but none of the colour or greyscale
functions are registered with the Lua state as of yet. All that needs to be
added is to add wrapper functions in the TC_Lua_Functions::Greyscale and
TC_Lua_Functions::RGB namespaces in TCAnimLua.cpp, and register these functions
with the RegisterCommands function in the respective namespace.

The DrawCube method in render.cpp was also heavily modified to support this
change. Each different type of animation is rendered in a different (but
fundamentally similar) loop for speed purposes.