Modified the current view rotation in the x-direction to wrap at +- 180 degrees. Also modified behaviour of the escape key (since it's close to the console key) so you have to press it twice within 0.5 seconds to quit the program.
Also, a new type called vectStr was defined in console_commands.cpp to avoid
typing std::vector<std::string> every time you want to create a vector of
strings (which is required for every command). Lastly, a new command called
"quit" was added, which provides the same functionality as the escape key.