Added a command queue, and the ability to insert a wait command between console commands. This should allow for advanced scripts to be loaded, giving the user the ability to wait for a certain amount of anything (time, ticks, iterations) before performing the next command. All further commands are sent to the commandQueue, and executed within the main event loop (pre-scene render) if and only if the current wait condition was met.
Also added was the ability to stop drawing the cube. Since the command queue
runs as part of the main render/event loop, if a user's framerate drops below
the current tick rate, the delay values may become inaccurate. If this happens,
the user can stop rendering the cube to raise their framerate to an acceptable
amount.