Added all basic required functionality for a given driver object, including the ability to run in a seperate thread (if asynchronous). To load or unload a driver, call the SetDriver function with the TCDriver object or NULL, respectively.
The documentation does need some significant updates, as much of it was simply
copied and pasted from the other existing files (to maintain a consistent
look). Finally, added header guards to TCDriver.h, and updated the ones in
TCAnim.h to maintain a more consistent naming convention.  Also removed the
threadInit variable from main.cpp (was only set, not used).