Cube list and connections now work with the 8x8x8 cube. Now all that has to be done is to add support for sending commands, and connecting manually to a cube.
Also, I encountered a strange bug - when the TCDriver_netdrv class' destructor
is called, the SDLNet_FreePacket call causes an access violation. The same
behaviour was noted when replacing the call with the delete operator. This
needs to be further investigated, but to avoid an error message every time the
program is closed, I changed udpPkt to be a static member of the
TCDriver_netdrv class, and it is only allocated once.